Экран профиля : Refactoring profile feature
