+++ app/src/main/java/ru/sr/mango_test_task/feature/profile/data/api/ProfileApi.kt
@@ -2,7 +2,6 @@ package ru.sr.mango_test_task.feature.profile.data.api
 
 import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.POST
 import retrofit2.http.PUT
 import ru.sr.mango_test_task.feature.profile.data.body.UpdateUserBody
 import ru.sr.mango_test_task.feature.profile.data.dto.ResponseCurrentUserDto
