+++ app/src/main/java/ru/sr/mango_test_task/feature/profile/data/dto/AvatarsDto.kt
@@ -1,7 +1,5 @@
 package ru.sr.mango_test_task.feature.profile.data.dto
 
-data class AvatarsDto(
-    val avatar: String,
-    val bigAvatar: String,
-    val miniAvatar: String
+class AvatarsDto(
+    val avatar: String
 )
