+++ app/src/main/java/ru/sr/mango_test_task/feature/profile/data/dto/ResponseUpdate.kt
@@ -1,5 +1,5 @@
 package ru.sr.mango_test_task.feature.profile.data.dto
 
-data class ResponseUpdate(
+class ResponseUpdate(
     val avatars: AvatarsDto?
 )
