+++ app/src/main/java/ru/sr/mango_test_task/feature/profile/di/ProfileUseCaseModule.kt
@@ -6,12 +6,12 @@ import ru.sr.mango_test_task.feature.profile.domain.repository.ProfileLocationRe
 import ru.sr.mango_test_task.feature.profile.domain.repository.ProfileRemoteRepository
 import ru.sr.mango_test_task.feature.profile.domain.usecase.GetUserUseCase
 import ru.sr.mango_test_task.feature.profile.domain.usecase.impl.GetUserUseCaseImpl
-import ru.sr.mango_test_task.feature.profile.domain.usecase.impl.UpdateUserUseCase
+import ru.sr.mango_test_task.feature.profile.domain.usecase.UpdateUserUseCase
 import ru.sr.mango_test_task.feature.profile.domain.usecase.impl.UpdateUserUseCaseImpl
 import javax.inject.Singleton
 
 @Module
-class ProfileUseCaseModule() {
+class ProfileUseCaseModule {
 
     @Provides
     @Singleton
