+++ app/src/main/java/ru/sr/mango_test_task/feature/profile/domain/usecase/GetUserUseCase.kt
@@ -3,6 +3,5 @@ package ru.sr.mango_test_task.feature.profile.domain.usecase
 import ru.sr.mango_test_task.feature.profile.domain.model.UserProfileDomainModel
 
 interface GetUserUseCase {
-
     suspend fun get(): UserProfileDomainModel
 }
