+++ app/src/main/java/ru/sr/mango_test_task/feature/profile/domain/usecase/impl/GetUserUseCaseImpl.kt
@@ -9,6 +9,7 @@ class GetUserUseCaseImpl(
     private val locationRepository: ProfileLocationRepository,
     private val remoteRepository: ProfileRemoteRepository,
 ) : GetUserUseCase {
+
     override suspend fun get(): UserProfileDomainModel {
         val locationUser = locationRepository.getCurrentUser()
         return if (locationUser != null) locationUser
