+++ app/src/main/java/ru/sr/mango_test_task/feature/profile/presentation/model/ProfileAction.kt
@@ -1,0 +1,5 @@
+package ru.sr.mango_test_task.feature.profile.presentation.model
+
+sealed interface ProfileAction {
+    object ShowSuccessToast : ProfileAction
+}
