+++ app/src/main/res/navigation/nav_graph.xml
@@ -54,7 +54,7 @@
     </fragment>
     <fragment
         android:id="@+id/profileFragment"
-        android:name="ru.sr.mango_test_task.feature.profile.ProfileFragment"
+        android:name="ru.sr.mango_test_task.feature.profile.presentation.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
